Add a second Entity related with foreign key. ManyToOne
-Copy previous entity project (with class, service, controller, and
repo) for simplicity
-Edit class, add @ManyToOne connection to PK attribute. Modify the
constructor to specify the PK relation Id.
-Edit Controller mapped request methods. Take care of /primaryKey/ as a
Path variable and modify each  request and methods associated to create,
update, delete and get the new entity with it foreign key association
--For GET ALL method, we have to custom the CrudRepository of the new
entity. Spring lets you manage and tweak the findBy.... method to any
property or attribute of the Entity!!! (Keep in mind to respect
CamelCase declaration)
--So in this case, we crate a method findByTopicId, where Topic is the
attribute and Id its primary key, to indicate to spring who to find de
related topic to that course. VERY USEFUL!!!